refactor before pull request
