+++ app/src/main/java/hu/unideb/smartcampus/main/activity/calendar/pojo/TimetableEvent.java
@@ -1,4 +1,4 @@
-package hu.unideb.smartcampus.fragment.calendar.pojo;
+package hu.unideb.smartcampus.main.activity.calendar.pojo;
 
 public class TimetableEvent {
